+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Slf4j now provides static methods to create and manage loggers
+### Changed
+- Slf4j Factory has been substituted with Context class
+
 ## [0.2.0] - 2017-02-26
 ### Added
 - Basic HTTP client